feat: 로그인 로직 유틸함수로 분리
- 재사용이 되지 않더라도 컴포넌트는 뷰와 상태에 집중하고 처리로직은 분리하기로 함

- #5
