+++ client/src/constants/env.ts
@@ -1,4 +1,4 @@
-export const baseUrl =
+export const BASE_URL =
   process.env.NODE_ENV === 'development' ? process.env.REACT_APP_CLIENT_DEV_URL : '';
 
-export const test = '';
+export const TEST = '';
