+++ client/src/pages/LoginPage.tsx
@@ -1,32 +1,10 @@
 import styled from 'styled-components';
 import BasicButton from '@components/common/BasicButton';
-import { baseUrl } from '@/constants/env';
+import { redirectToAuthUrl, OAuthOriginType } from '@/utils/oAuth';
 
-type OAuthOriginType = 'GITHUB' | 'NAVER';
-
-enum AuthUrlEnum {
-  GITHUB = 'https://github.com/login/oauth/authorize',
-  NAVER = '',
-}
 export default function LoginPage() {
   const handleClickOauthButton = async (oAuthOrigin: OAuthOriginType) => {
-    const authUrl = AuthUrlEnum[oAuthOrigin];
-
-    if (!process.env.REACT_APP_CLIENT_ID) {
-      throw new Error('Cannot find client id');
-    }
-
-    const queryConfig = {
-      scope: 'read:user',
-      client_id: process.env.REACT_APP_CLIENT_ID,
-      redirect_uri: `${baseUrl}/oauth-redirect`,
-    };
-
-    const searchParamsObj = new URLSearchParams(queryConfig);
-    const queryString = `?${searchParamsObj.toString()}`;
-
-    const githubLoginUrl = authUrl + queryString;
-    window.location.href = githubLoginUrl;
+    redirectToAuthUrl(oAuthOrigin);
   };
 
   return (
